+++ nuxeo-core/nuxeo-core/src/main/java/org/nuxeo/ecm/core/blob/BlobManager.java
@@ -35,43 +35,6 @@ import org.nuxeo.ecm.core.api.Blob;
 public interface BlobManager {
 
     /**
-     * Class describing information from a {@link Blob}, suitable for serialization and storage.
-     *
-     * @since 7.2
-     */
-    class BlobInfo {
-        public String key;
-
-        public String mimeType;
-
-        public String encoding;
-
-        public String filename;
-
-        public Long length;
-
-        public String digest;
-
-        /** Empty constructor. */
-        public BlobInfo() {
-        }
-
-        /**
-         * Copy constructor.
-         *
-         * @since 7.10
-         */
-        public BlobInfo(BlobInfo other) {
-            key = other.key;
-            mimeType = other.mimeType;
-            encoding = other.encoding;
-            filename = other.filename;
-            length = other.length;
-            digest = other.digest;
-        }
-    }
-
-    /**
      * Hints for returning {@link URI}s appropriate for the expected usage.
      *
      * @since 7.3
