+++ nuxeo-core/nuxeo-core/src/main/java/org/nuxeo/ecm/core/blob/FilesystemBlobProvider.java
@@ -30,7 +30,6 @@ import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.lang.StringUtils;
 import org.nuxeo.ecm.core.api.Blob;
 import org.nuxeo.ecm.core.api.NuxeoException;
-import org.nuxeo.ecm.core.blob.BlobManager.BlobInfo;
 
 /**
  * Blob provider that can reference files on the filesystem.
